♻️ refactor: database
