+++ config/factory_model_map.php
@@ -7,11 +7,11 @@ return [
     | Used when models live in subnamespaces (e.g. App\Models\Catalog\Item).
     |--------------------------------------------------------------------------
     */
-    \App\Models\Catalog\Item::class => \Database\Factories\ItemFactory::class,
-    \App\Models\Catalog\Section::class => \Database\Factories\SectionFactory::class,
-    \App\Models\Catalog\ItemComponent::class => \Database\Factories\ItemComponentFactory::class,
-    \App\Models\Catalog\Extra::class => \Database\Factories\ExtraFactory::class,
-    \App\Models\Catalog\TagItem::class => \Database\Factories\TagItemFactory::class,
-    \App\Models\Taxonomy\Tag::class => \Database\Factories\TagFactory::class,
-    \App\Models\Proprietary\Proprietary::class => \Database\Factories\ProprietaryFactory::class,
+    \App\Models\Catalog\Item::class => \Database\Factories\Catalog\ItemFactory::class,
+    \App\Models\Catalog\Section::class => \Database\Factories\Catalog\SectionFactory::class,
+    \App\Models\Catalog\ItemComponent::class => \Database\Factories\Catalog\ItemComponentFactory::class,
+    \App\Models\Catalog\Extra::class => \Database\Factories\Catalog\ExtraFactory::class,
+    \App\Models\Catalog\TagItem::class => \Database\Factories\Taxonomy\TagItemFactory::class,
+    \App\Models\Taxonomy\Tag::class => \Database\Factories\Taxonomy\TagFactory::class,
+    \App\Models\Proprietary\Proprietary::class => \Database\Factories\Proprietary\ProprietaryFactory::class,
 ];
