+++ database/factories/Catalog/SectionFactory.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Database\Factories;
+namespace Database\Factories\Catalog;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
