+++ database/factories/Proprietary/ProprietaryFactory.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Database\Factories;
+namespace Database\Factories\Proprietary;
 
 use Illuminate\Database\Eloquent\Factories\Factory;
